requirements.txt
Django==4.1.5
psycopg2
Pillow
pytz
# gdal must match the gdal libraries on OS, this is gdal on bullseye
GDAL==3.2.2.1
simplejson
numpy
